Moving Controller to Separate File
