+++ src/main/kotlin/distributed/lock/LockApplication.kt
@@ -14,15 +14,3 @@ class LockApplication
 fun main(args: Array<String>) {
 	runApplication<LockApplication>(*args)
 }
-
-@RequestMapping("/dt")
-@RestController()
-class MainController {
-
-
-	@DistributedLock(clazz = MainController::class)
-	@GetMapping("/test/{v1}/{v2}/{v3}")
-	fun test(@PathVariable v1: String, @PathVariable v2: String, @PathVariable @KeyVariable v3 : String) : String {
-		return "Check"
-	}
-}
